blabla

+++ includes/Node.hpp
@@ -9,7 +9,7 @@ class Node
 	 	uint16_t 			_indexChild;
 
 	public:
-		uint8_t 			depth;
+		int					depth;
 		// int16_t				lowerBound;
 		// int16_t				upperBound;
 		bool 				maxNode;
@@ -20,7 +20,7 @@ class Node
 		Node*				getNextChild();
 		void 				insertChild(Node* node);
 
-		Node(uint8_t map[SIZE_BOARD][SIZE_BOARD], uint16_t moveCoord, bool maxNode, uint8_t depth, uint8_t player);
+		Node(uint8_t map[SIZE_BOARD][SIZE_BOARD], int16_t moveCoord, bool maxNode, int depth, uint8_t player);
 		~Node();
 	
 };
